+++ mono/mini/il2cpp-stubs.cpp
@@ -1482,7 +1482,7 @@ MonoClass* il2cpp_mono_get_string_class (void)
 	return (MonoClass*)il2cpp_defaults.string_class;
 }
 
-Il2CppSequencePoint* il2cpp_get_sequence_point(size_t id)
+Il2CppSequencePoint* il2cpp_get_sequence_point(int id)
 {
 #if IL2CPP_MONO_DEBUGGER
     return il2cpp::utils::Debugger::GetSequencePoint(id);
